requirements.txt
# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.3

# Data Processing
pandas>=2.2.0

# LLM
google-generativeai==0.3.2

# Sentiment Analysis
transformers>=4.36.0
torch>=2.1.0

# UI
streamlit>=1.40.0

# Visualization
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0

# Utilities
python-dotenv==1.0.0

# RAG & Vector Store
langchain==0.1.0
langchain-google-genai==0.0.5
chromadb==0.4.22
sentence-transformers==2.3.1
